requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-chroma==0.1.4
langchain-google-genai==2.0.0
langchain-huggingface==0.1.0
sentence-transformers==3.1.0
rank_bm25==0.2.2
python-docx==1.1.2
python-dotenv==1.0.1
regex